Voltei atrás com o método getAthletesResults, deixei na classe User mesmo
